+++ packages/cli-hydrogen/src/cli/commands/hydrogen/dev.ts
@@ -1,4 +1,5 @@
 import devService from '../../services/dev'
+import {hydrogenFlags} from '../../flags'
 import {path, cli} from '@shopify/cli-kit'
 import {Command, Flags} from '@oclif/core'
 
@@ -6,11 +7,7 @@ export default class Dev extends Command {
   static description = 'Run a Hydrogen storefront locally for development'
   static flags = {
     ...cli.globalFlags,
-    path: Flags.string({
-      hidden: true,
-      description: 'the path to your hydrogen storefront',
-      env: 'SHOPIFY_FLAG_PATH',
-    }),
+    path: hydrogenFlags.path,
     force: Flags.boolean({
       description: 'force dependency pre-bundling.',
       env: 'SHOPIFY_FLAG_DEV_FORCE',