+++ packages/create-app/src/prompts/init.test.ts
@@ -6,7 +6,7 @@ describe('init', () => {
     const prompt = vi.fn()
     const answers = {
       name: 'app',
-      template: 'https://github.com/Shopify/starter-node-app#add-shopify-home-toml',
+      template: 'https://github.com/Shopify/shopify-app-template-node#add-shopify-home-toml',
     }
     const options = {template: 'template'}
 
@@ -23,7 +23,7 @@ describe('init', () => {
         name: 'name',
         preface: 'Welcome. Let’s get started by naming your app. You can change it later.',
         message: "Your app's name?",
-        default: 'app',
+        default: expect.stringMatching(/^\w+-\w+-app-\d+$/),
         validate: expect.any(Function),
       },
     ])
@@ -34,7 +34,7 @@ describe('init', () => {
     const prompt = vi.fn()
     const answers = {
       name: 'app',
-      template: 'https://github.com/Shopify/starter-node-app#add-shopify-home-toml',
+      template: 'https://github.com/Shopify/shopify-app-template-node#add-shopify-home-toml',
     }
     const options = {name: 'app', template: 'template'}